(issue #50) add baseclasses into one single file
